+++ android/src/main/kotlin/net/mullvad/mullvadvpn/util/ListenableScrollableView.kt
@@ -1,0 +1,8 @@
+package net.mullvad.mullvadvpn.util
+
+interface ListenableScrollableView {
+    val horizontalScrollOffset: Int
+    val verticalScrollOffset: Int
+
+    var onScrollListener: ((Int, Int, Int, Int) -> Unit)?
+}